One Front Door/SourceCode/DPSG.Web.Merchandiser/src/WebApp/src/app/dispatch/image-list/image-list.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { ImageListComponent } from './image-list.component';

describe('Component: ImageList', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [ImageListComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([ImageListComponent],
      (component: ImageListComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(ImageListComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(ImageListComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-image-list></app-image-list>
  `,
  directives: [ImageListComponent]
})
class ImageListComponentTestController {
}

